Fix Kalshi markets endpoint type mismatch and enhance regression tests
- Fix type mismatch in Kalshi markets endpoint (i64 -> i32 for limit/offset)
- Update regression script with comprehensive new tests:
  - Additional API endpoints (market status, news, Kalshi, FRED)
  - Container health and resource monitoring
  - Database performance and connection pool checks
  - API response time monitoring
  - Data integrity and consistency validation
  - Network connectivity verification
  - Environment variable validation
- Fix Xcode scheme path in regression script
- Improve network connectivity tests to use health endpoints

+++ capstones/therealtplum/apps/rust-api/src/kalshi.rs
@@ -81,6 +81,7 @@ pub async fn list_kalshi_markets_handler(
     let search = params.search.as_deref();
 
     // Use the database function for filtering
+    // Note: Cast limit and offset to i32 to match database function signature (integer type)
     let result = sqlx::query_as::<_, KalshiMarketSummary>(
         r#"
         SELECT * FROM get_kalshi_markets_filtered($1, $2, $3, $4, $5)
@@ -88,8 +89,8 @@ pub async fn list_kalshi_markets_handler(
     )
     .bind(category)
     .bind(status)
-    .bind(limit)
-    .bind(offset)
+    .bind(limit as i32)
+    .bind(offset as i32)
     .bind(search)
     .fetch_all(&state.db_pool)
     .await;
